[Refactor] utils 파일 및 헤더 정리 #58
