+++ srcs/execute/builtin.c
@@ -1,4 +1,5 @@
 #include "../../includes/execute.h"
+#include <unistd.h>
 
 int	is_builtin(char **cmd)
 {
