+++ srcs/execute/builtin_echo.c
@@ -1,4 +1,5 @@
 #include "../../includes/execute.h"
+#include <unistd.h>
 
 static void	change_flag(char ***cmd, int *flag, char **tmp, int i)
 {
