+++ srcs/execute/builtin_exit.c
@@ -1,4 +1,5 @@
 #include "../../includes/execute.h"
+#include <unistd.h>
 
 static int	ft_atouc(char *str)
 {
