+++ srcs/execute/builtin_pwd.c
@@ -1,5 +1,6 @@
 #include "../../includes/execute.h"
 #include <limits.h>
+#include <unistd.h>
 
 void	ft_pwd(char **cmd)
 {
