+++ srcs/execute/heredoc.c
@@ -1,4 +1,6 @@
 #include "../../includes/execute.h"
+#include <readline/readline.h>
+#include <unistd.h>
 
 void	unlink_heredoc(t_argv *argv)
 {
