+++ srcs/execute/open.c
@@ -1,5 +1,6 @@
 #include "../../includes/execute.h"
 #include <fcntl.h>
+#include <unistd.h>
 
 int	ft_open(char *file, int type)
 {
