+++ srcs/execute/pipe.c
@@ -1,4 +1,5 @@
 #include "../../includes/execute.h"
+#include <unistd.h>
 
 void	set_stdin_pipe(int **pipes, int num)
 {
