+++ srcs/execute/redir.c
@@ -1,4 +1,5 @@
 #include "../../includes/execute.h"
+#include <unistd.h>
 
 void	reset_stdin(int fd)
 {
