+++ srcs/utils/env1.c
@@ -1,4 +1,5 @@
 #include "../../includes/utils.h"
+#include <string.h>
 #include <errno.h>
 
 t_env	*get_env(char *key)
