+++ srcs/utils/env2.c
@@ -1,4 +1,5 @@
 #include "../../includes/utils.h"
+#include <string.h>
 #include <errno.h>
 
 void	make_env_arr(void)
